requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
PyMySQL==1.1.1
PyYAML==6.0.2
pydantic==2.11.3
redis==6.0.0
openpyxl==3.1.5
